chore: move fetch-docs to scripts/utils/migrate
